docs: record why, and tighten the build to -Wconversion
Adds docs/adr/, seven records covering the decisions this migration
made: world storage and why not SoA, page rendering and why not one
texture per chunk, the active-chunk schedule and why the tick's set is
frozen, the event buffer and why not a bus, the deterministic RNG and
why separate streams, the ability registry and why not data-driven, and
the lighting window. Each one names what was rejected and on what
measurement — the channel-split lighting experiment in particular, so it
does not get "improved" back.

Adds docs/development/adding-an-entity.md. There is no entity system
yet, and the document says so; what it is for is the rule that must not
be broken when one arrives — the fourteen million world cells are not
entities, and an ECS gets introduced when something concrete needs it,
not because large engines have one.

Turns on -Wconversion. It found exactly three implicit int-to-float
conversions across app, tests and benchmark, all fixed, and produces no
noise, so it is worth keeping permanently rather than the usual verdict
on that flag.

Adds a compile_commands.json target for clangd, generated from the same
source lists the build uses, ignored by git and removed by make clean.

+++ src/particles.c
@@ -297,8 +297,10 @@ void ParticlesSpawnBoostTrail(ParticleSystem *system, Vector2 position,
                                     position.y + tangent.y * spread},
                           (Vector2){-direction.x * backward + tangent.x * spread,
                                     -direction.y * backward + tangent.y * spread},
-                          color, 0.12f + RandomUnit(system) * (0.12f + stage * 0.04f),
-                          0.35f + RandomUnit(system) * (0.45f + stage * 0.22f), 0.0f);
+                          color,
+                          0.12f + RandomUnit(system) * (0.12f + (float)stage * 0.04f),
+                          0.35f + RandomUnit(system) * (0.45f + (float)stage * 0.22f),
+                          0.0f);
     }
 }
